+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -7,7 +7,7 @@ import com.techcourse.controller.RegisterViewController;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.controller.asis.ForwardController;
 import org.slf4j.Logger;
