+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -7,7 +7,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import nextstep.mvc.Handleradaptor.HandlerAdapter;
+import nextstep.mvc.handleradaptor.HandlerAdapter;
+import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
