+++ mvc/src/main/java/nextstep/mvc/handleradaptor/HandlerExecutionHandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.Handleradaptor;
+package nextstep.mvc.handleradaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,7 +6,7 @@ import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class HandlerExecutionAdapter implements HandlerAdapter {
+public class HandlerExecutionHandlerAdapter implements HandlerAdapter {
     @Override
     public boolean supports(Object handler) {
         return handler instanceof HandlerExecution;
