+++ mvc/src/main/java/nextstep/mvc/handleradaptor/ManualHandlerAdaptor.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.Handleradaptor;
+package nextstep.mvc.handleradaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
